@@ -15,8 +15,8 @@
     "@reduxjs/toolkit": "^1.9.7",
     "axios": "^1.6.0",
     "classnames": "^2.3.2",
-    "next": "13.5",
-    "next-redux-wrapper": "^8.1.0",
+    "next": "^14.0.3",
+    "next-auth": "^4.24.5",
     "react": "^18",
     "react-dom": "^18",
     "react-loader-spinner": "^5.4.5",
