@@ -1,65 +1,34 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import React from 'react';
 import axios from 'axios';
-import { useDispatch } from 'react-redux';
-import { useCheckAuth } from '../auth/useCheckAuth';
 import { url } from '@/api';
 import { setUser } from '@/redux/slices/authSlice';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
+import { useSession } from 'next-auth/react';
 
-export const useOrder = () => {
-  const accessToken = localStorage.getItem('accessToken');
+export const useOrder = (orderId: any) => {
+  const { data: session }: any = useSession();
   const router = useRouter();
-  const { id } = router.query;
   const [order, setOrder] = React.useState<any>();
   const [orderError, setOrderError] = React.useState('');
   const [orderLoading, setOrderLoading] = React.useState(false);
-  const { refresh } = useCheckAuth();
-  const dispatch = useDispatch();
 
   React.useEffect(() => {
     setOrderLoading(true);
-    axios.get(`${url}/orders/?id=${id}`, {
+    axios.get(`${url}/orders/?id=${orderId}`, {
       headers: {
-        Authorization: `Bearer ${accessToken}`,
+        Authorization: `Bearer ${session?.user?.accessToken}`,
       },
     })
       .then(response => {
         setOrder(response.data);
       })
       .catch( (e) => {
-        if (e.response.status === 401) {
-          refresh()
-            .then(() => {
-              const accessToken = localStorage.getItem('accessToken');
-              axios.get(`${url}/orders/?id=${id}`, {
-                headers: {
-                  Authorization: `Bearer ${accessToken}`,
-                },
-                withCredentials: true,
-              })
-                .then(response => {
-                  setOrder(response.data);
-                  setOrderLoading(true);
-                })
-                .catch((e) => {
-                  if (e.response.status === 401) {
-                    dispatch(setUser(null));
-                    localStorage.setItem('accessToken', '');
-                  }
-                  setOrderError(e.response.data.message);
-                  setOrderLoading(false);
-                })
-                .finally(() => {
-                  setOrderLoading(false);
-                });
-            });
-        }
       })
       .finally(() => {
         setOrderLoading(false);
       });
-  }, []);
+  }, [session]);
 
   return { order, orderError, orderLoading };
 };
