@@ -3,23 +3,16 @@ import { useOrders } from '@/api/services/orders/useOrders';
 import React from 'react'
 import { Loading } from '../ui/loading/loading';
 import Link from 'next/link';
-import { useSelector } from 'react-redux';
-import { useRouter } from 'next/navigation';
-import { RootState } from '@/redux/store';
+import { Provider } from 'react-redux';
+import { store } from '@/redux/store';
 
 
 export const AcoountPersonal = () => {
   const { ordersNormalized, ordersError, ordersLoading } = useOrders();
-  const { user } = useSelector((state: RootState) => state.auth);
-  const router = useRouter();
-
-  if (!user) {
-    router.replace('/login');
-    return;
-  }
 
   return (
-    <div>
+    <Provider store={store}>
+      <div>
       <h5 className='personalAccount__ordersTitle'>
         {ordersNormalized && ordersNormalized.length > 0 
           ? 'Ваші замовлення' : ordersLoading === true ? 'Завантаження' : 'Замовлень немає'}
@@ -29,12 +22,13 @@ export const AcoountPersonal = () => {
       ) : (
         <div>
           {ordersNormalized && ordersNormalized.map((order: any) => (
-            <Link href={`orders/${order.id.toString()}`} key={order.id}>
+            <Link href={`account/orders/${order.id.toString()}`} key={order.id}>
               <p>Дата замовлення: {order.createdAt.toUTCString().slice(0, -4)}</p>
             </Link>
           ))}
         </div>
       )}
     </div>
+    </Provider>
   )
 }
